% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt_ppp.R
\name{cmt_to_ppp}
\alias{cmt_to_ppp}
\title{Title}
\usage{
cmt_to_ppp(cmt = NA, limsx = c(-180, 180), limsy = c(-90, 90),
  ppp = T, simplify = F)
}
\arguments{
\item{cmt}{}

\item{limsx}{}

\item{limsy}{}

\item{ppp}{}

\item{simplify}{}
}
\description{
Title
}
