% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt.R
\name{filter_depth}
\alias{filter_depth}
\title{filter by depth}
\usage{
filter_depth(f = c(0, 300), data = NA, quick = F)
}
\arguments{
\item{f}{default c(),300)}

\item{data}{default=NA in which case cmt.RDS of qmt.RDS is read}

\item{quick}{dafault=F, set to TRUE if quick cmt is requird.}
}
\value{
filter cmt tibble
}
\description{
filter by depth
}
