% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt.R
\name{filter_lat}
\alias{filter_lat}
\title{filter by lat range}
\usage{
filter_lat(f = c(-25, 10), data = NA, quick = F)
}
\arguments{
\item{f}{(default c(-25,10))}

\item{data}{(default = NA) will call read_cmtRDS}

\item{quick}{}
}
\value{

}
\description{
filter by lat range
}
