% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt.R
\name{filter_long}
\alias{filter_long}
\title{filter by long range}
\usage{
filter_long(f = c(100, 160), data = NA, quick = F)
}
\arguments{
\item{f}{(default c(100,160))}

\item{data}{(default = NA) will call read_cmtRDS}

\item{quick}{}
}
\value{

}
\description{
filter by long range
}
