% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt.R
\name{filter_mw}
\alias{filter_mw}
\title{filter by mw}
\usage{
filter_mw(f = 7, data = NA, quick = F)
}
\arguments{
\item{f}{defult f=7}

\item{data}{default=NA in whcihc ase cmt.RDS of qmt.RDS is read}

\item{quick}{dafault=F, set to TRUE if quick cmt is requird.}
}
\value{
filter cmt tibble
}
\description{
filter by mw
}
