% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt.R
\name{filter_region}
\alias{filter_region}
\title{filter by geographic region name in regex}
\usage{
filter_region(data = NA, f = "TIMOR|BANDA|INDONE", quick = F)
}
\arguments{
\item{data}{defaults to "TIMOR|BANDA|INDONE"}

\item{f}{}

\item{quick}{}
}
\value{

}
\description{
filter by geographic region name in regex
}
