% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt.R
\name{minmax}
\alias{minmax}
\title{returns min and max of nuemric vector expanded by amount}
\usage{
minmax(data, tight = 0)
}
\arguments{
\item{data}{}

\item{tight}{(default=0)   factor in percentage terms to expand range}
}
\value{

}
\description{
returns min and max of nuemric vector expanded by amount
}
