% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnt_line}
\alias{pnt_line}
\title{projection of point onto line}
\usage{
pnt_line(p0, l0, l1)
}
\arguments{
\item{p0}{A numeric (length 2)  vector of the point ( long lat)  to projected onto the line defined by l0, l1}

\item{l0}{A numeric (length 2) vector of the point (long lat) defining start of projection line}

\item{l1}{A numeric (length 2) vector of the point (long lat) defining end of projection line}
}
\description{
This function returns projection point
}
