% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnt_line_vec}
\alias{pnt_line_vec}
\title{projection of points onto line}
\usage{
pnt_line_vec(p0, l0, l1, profile)
}
\arguments{
\item{p0}{A nuemric matrix (n*2) defining points in (long lat)  format to be projected onto line defined by l0, l1}

\item{l0}{A numeric vector}

\item{l1}{A numeric vector}

\item{profiel}{A integer identfier for the profile}
}
\description{
This function returns projection point
}
