% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{project_ppp}
\alias{project_ppp}
\title{Title}
\usage{
project_ppp(
  my.ppp,
  win = pwin(my.ppp),
  plates = bird.plates.psp,
  profile = "p0",
  add = F,
  test = F
)
}
\arguments{
\item{my.ppp}{}

\item{win}{}

\item{plates}{}

\item{profile}{}

\item{add}{}

\item{test}{}
}
\value{

}
\description{
Title
}
