% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt.R
\name{read_cmtRDS}
\alias{read_cmtRDS}
\title{read .RDS file in cmt package}
\usage{
read_cmtRDS(quick = F)
}
\arguments{
\item{quick}{(default= FALSE) set to TRUE to read quick cmt RDS}
}
\value{
list of 2,  data (class tibble)  and date (class Date)
}
\description{
read .RDS file in cmt package
}
