% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt.R
\name{set_cmt_sys}
\alias{set_cmt_sys}
\title{set_cmt_sys sets a lcoal directory for the}
\usage{
set_cmt_sys(dir = NA)
}
\arguments{
\item{dir}{defaults to "~/data/global/quakes/cmt17/" recursively genartes this if not in existence}
}
\value{

}
\description{
set_cmt_sys sets a lcoal directory for the
}
