% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt.R
\name{vec_lines}
\alias{vec_lines}
\title{return trend lines for cmt P, T or B axes}
\usage{
vec_lines(
  cmt,
  scale = 1,
  psp = F,
  segment = F,
  norm = T,
  simplify = T,
  ax = "P"
)
}
\arguments{
\item{cmt}{}

\item{scale}{}

\item{psp}{}

\item{segment}{}

\item{norm}{}

\item{simplify}{}

\item{ax}{(default="P")}
}
\value{

}
\description{
return trend lines for cmt P, T or B axes
}
